python-service/requirements.txt
fastapi
uvicorn
faiss-cpu
sentence-transformers
pandas
numpy
python-dotenv
groq
tavily-python
langgraph
langchain-core